+++ R/bayes_mcmc.R
@@ -2,7 +2,7 @@
 #'
 #' Perform MCMC JAGS sampling or HMC Stan sampling for Gaussian mixture models, post-process the chains and apply a clustering technique to the MCMC sample. Pivotal units for each group are selected among four alternative criteria.
 #' @param y \eqn{N}-dimensional vector for univariate data or
-#' \eqn{N \times 2} matrix for bivariate data.
+#' \eqn{N \times D} matrix for bivariate data.
 #' @param k Number of mixture components.
 #' @param nMC Number of MCMC iterations for the JAGS/Stan function execution.
 #' @param priors Input prior hyperparameters (see Details for default options).
@@ -22,33 +22,33 @@
 #' \code{software="rstan"}). Default is 1.
 #'
 #' @details
-#' The function fits univariate and bivariate Bayesian Gaussian mixture models of the form
+#' The function fits univariate and multivariate Bayesian Gaussian mixture models of the form
 #' (here for univariate only):
-#' \deqn{(Y_i|Z_i=j) \sim \mathcal{N}(\mu_j,\phi_j),}
+#' \deqn{(Y_i|Z_i=j) \sim \mathcal{N}(\mu_j,\sigma_j),}
 #' where the \eqn{Z_i}, \eqn{i=1,\ldots,N}, are i.i.d. random variables, \eqn{j=1,\dots,k},
-#' \eqn{\phi_j} is the group variance,  \eqn{Z_i \in {1,\ldots,k }} are the
+#' \eqn{\sigma_j} is the group variance,  \eqn{Z_i \in {1,\ldots,k }} are the
 #' latent group allocation, and
-#' \deqn{P(Z_i=j)=\pi_j.}
+#' \deqn{P(Z_i=j)=\eta_j.}
 #' The likelihood of the model is then
-#' \deqn{L(y;\mu,\pi,\phi) = \prod_{i=1}^N \sum_{j=1}^k \pi_j \mathcal{N}(\mu_j,\phi_j),}
-#' where \eqn{(\mu, \phi)=(\mu_{1},\dots,\mu_{k},\phi_{1},\ldots,\phi_{k})}
-#' are the component-specific parameters and \eqn{\pi=(\pi_{1},\dots,\pi_{k})}
+#' \deqn{L(y;\mu,\eta,\sigma) = \prod_{i=1}^N \sum_{j=1}^k \eta_j \mathcal{N}(\mu_j,\sigma_j),}
+#' where \eqn{(\mu, \sigma)=(\mu_{1},\dots,\mu_{k},\sigma_{1},\ldots,\sigma_{k})}
+#' are the component-specific parameters and \eqn{\eta=(\eta_{1},\dots,\eta_{k})}
 #' the mixture weights. Let \eqn{\nu} denote a permutation of \eqn{{ 1,\ldots,k }},
 #' and let \eqn{\nu(\mu)= (\mu_{\nu(1)},\ldots,} \eqn{ \mu_{\nu(k)})},
-#' \eqn{\nu(\phi)= (\phi_{\nu(1)},\ldots,} \eqn{ \phi_{\nu(k)})},
-#' \eqn{ \nu(\pi)=(\pi_{\nu(1)},\ldots,\pi_{\nu(k)})} be the
-#' corresponding permutations of \eqn{\mu}, \eqn{\phi} and \eqn{\pi}.
+#' \eqn{\nu(\sigma)= (\sigma_{\nu(1)},\ldots,} \eqn{ \sigma_{\nu(k)})},
+#' \eqn{ \nu(\eta)=(\eta_{\nu(1)},\ldots,\eta_{\nu(k)})} be the
+#' corresponding permutations of \eqn{\mu}, \eqn{\sigma} and \eqn{\eta}.
 #'  Denote by \eqn{V} the set of all the permutations of the indexes
 #'  \eqn{{1,\ldots,k }}, the likelihood above is invariant under any
 #'  permutation \eqn{\nu \in V}, that is
 #' \deqn{
-#' L(y;\mu,\pi,\phi) = L(y;\nu(\mu),\nu(\pi),\nu(\phi)).}
+#' L(y;\mu,\eta,\sigma) = L(y;\nu(\mu),\nu(\eta),\nu(\sigma)).}
 #' As a consequence, the model is unidentified with respect to an
 #' arbitrary permutation of the labels.
 #' When Bayesian inference for the model is performed,
-#' if the prior distribution \eqn{p_0(\mu,\pi,\phi)} is invariant under a permutation of the indices, then so is the posterior. That is, if \eqn{p_0(\mu,\pi,\phi) = p_0(\nu(\mu),\nu(\pi),\phi)}, then
+#' if the prior distribution \eqn{p_0(\mu,\eta,\sigma)} is invariant under a permutation of the indices, then so is the posterior. That is, if \eqn{p_0(\mu,\eta,\sigma) = p_0(\nu(\mu),\nu(\eta),\sigma)}, then
 #'\deqn{
-#' p(\mu,\pi,\phi| y) \propto p_0(\mu,\pi,\phi)L(y;\mu,\pi,\phi)}
+#' p(\mu,\eta,\sigma| y) \propto p_0(\mu,\eta,\sigma)L(y;\mu,\eta,\sigma)}
 #' is multimodal with (at least) \eqn{k!} modes.
 #'
 #' Depending on the selected software, the model parametrization
@@ -63,8 +63,8 @@
 #' \code{bayesmix} package:
 #'
 #'  \deqn{\mu_j \sim \mathcal{N}(\mu_0, 1/B0inv)}
-#'  \deqn{\phi_j \sim \mbox{invGamma}(nu0Half, nu0S0Half)}
-#'  \deqn{\pi \sim \mbox{Dirichlet}(1,\ldots,1)}
+#'  \deqn{\sigma_j \sim \mbox{invGamma}(nu0Half, nu0S0Half)}
+#'  \deqn{\eta \sim \mbox{Dirichlet}(1,\ldots,1)}
 #'  \deqn{S0 \sim \mbox{Gamma}(g0Half, g0G0Half),}
 #'
 #'  with default values: \eqn{\mu_0=0, B0inv=0.1, nu0Half =10, S0=2,
@@ -80,19 +80,19 @@
 #'  When \code{software="rstan"}, the prior specification is:
 #'
 #'  \deqn{\mu_j \sim \mathcal{N}(\mu_0, 1/B0inv)}
-#'  \deqn{\phi_j \sim \mbox{Lognormal}(\mu_{\phi}, \sigma_{\phi})}
-#'  \deqn{\pi_j \sim \mbox{Uniform}(0,1),}
+#'  \deqn{\sigma_j \sim \mbox{Lognormal}(\mu_{\sigma}, \tau_{\sigma})}
+#'  \deqn{\eta_j \sim \mbox{Uniform}(0,1),}
 #'
-#'  with default values: \eqn{\mu_0=0, B0inv=0.1, \mu_{\phi}=0, \sigma_{\phi}=2}.
+#'  with default values: \eqn{\mu_0=0, B0inv=0.1, \mu_{\sigma}=0, \tau_{\sigma}=2}.
 #' The users may specify new hyperparameter values with the argument:
 #'
-#' \code{priors=list(mu_0=1, B0inv=0.2, mu_phi=3, sigma_phi=5)}
+#' \code{priors=list(mu_0=1, B0inv=0.2, mu_sigma=3, tau_sigma=5)}
 #'
-#'For bivariate mixtures, when \code{software="rjags"} the prior specification is the following:
+#'For multivariate mixtures, when \code{software="rjags"} the prior specification is the following:
 #'
-#'\deqn{ \bm{\mu}_j  \sim \mathcal{N}_2(\bm{\mu}_0, S2)}
-#'\deqn{ 1/\Sigma \sim \mbox{Wishart}(S3, 3)}
-#'\deqn{\pi \sim \mbox{Dirichlet}(\bm{\alpha}),}
+#'\deqn{ \bm{\mu}_j  \sim \mathcal{N}_D(\bm{\mu}_0, S2)}
+#'\deqn{ \Sigma^{-1} \sim \mbox{Wishart}(S3, D+1)}
+#'\deqn{\eta \sim \mbox{Dirichlet}(\bm{\alpha}),}
 #'
 #'where  \eqn{\bm{\alpha}} is a \eqn{k}-dimensional vector
 #'and \eqn{S_2} and \eqn{S_3}
@@ -110,27 +110,27 @@
 #'
 #'When \code{software="rstan"}, the prior specification is:
 #'
-#'\deqn{ \bm{\mu}_j  \sim \mathcal{N}_2(\bm{\mu}_0, LDL^{T})}
-#'\deqn{L \sim \mbox{LKJ}(\eta)}
-#'\deqn{D_j \sim \mbox{HalfCauchy}(0, \sigma_d).}
+#'\deqn{ \bm{\mu}_j  \sim \mathcal{N}_D(\bm{\mu}_0, LD*L^{T})}
+#'\deqn{L \sim \mbox{LKJ}(\epsilon)}
+#'\deqn{D^*_j \sim \mbox{HalfCauchy}(0, \sigma_d).}
 #'
-#'The covariance matrix is expressed in terms of the LDL decomposition as \eqn{LDL^{T}},
-#'a variant of the classical Cholesky decomposition, where \eqn{L} is a \eqn{2 \times 2}
-#'lower unit triangular matrix and \eqn{D} is a \eqn{2 \times 2} diagonal matrix.
-#'The Cholesky correlation factor \eqn{L} is assigned a LKJ prior with \eqn{\eta} degrees of freedom,  which,
+#'The covariance matrix is expressed in terms of the LDL decomposition as \eqn{LD*L^{T}},
+#'a variant of the classical Cholesky decomposition, where \eqn{L} is a \eqn{D \times D}
+#'lower unit triangular matrix and \eqn{D*} is a \eqn{D \times D} diagonal matrix.
+#'The Cholesky correlation factor \eqn{L} is assigned a LKJ prior with \eqn{\epsilon} degrees of freedom,  which,
 #'combined with priors on the standard deviations of each component, induces a prior on the covariance matrix;
-#'as \eqn{\eta \rightarrow \infty} the magnitude of correlations between components decreases,
-#'whereas \eqn{\eta=1} leads to a uniform prior distribution for \eqn{L}.
-#'By default, the hyperparameters are \eqn{\bm{\mu}_0=\bm{0}}, \eqn{\sigma_d=2.5, \eta=1}.
+#'as \eqn{\epsilon \rightarrow \infty} the magnitude of correlations between components decreases,
+#'whereas \eqn{\epsilon=1} leads to a uniform prior distribution for \eqn{L}.
+#'By default, the hyperparameters are \eqn{\bm{\mu}_0=\bm{0}}, \eqn{\sigma_d=2.5, \epsilon=1}.
 #'The user may propose some different values with the argument:
 #'
 #'
-#' \code{priors=list(mu_0=c(1,2), sigma_d = 4, eta =2)}
+#' \code{priors=list(mu_0=c(1,2), sigma_d = 4, epsilon =2)}
 #'
 #'
 #' If \code{software="rjags"} the function performs JAGS sampling using the \code{bayesmix} package
 #' for univariate Gaussian mixtures, and the \code{runjags}
-#' package for bivariate Gaussian mixtures. If \code{software="rstan"} the function performs
+#' package for multivariate Gaussian mixtures. If \code{software="rstan"} the function performs
 #' Hamiltonian Monte Carlo (HMC) sampling via the \code{rstan} package (see the vignette and the Stan project
 #' for any help).
 #'
@@ -162,21 +162,21 @@
 #' vector.}
 #' \item{\code{mcmc_mean}}{  If \code{y} is a vector, a \eqn{true.iter \times k}
 #' matrix with the post-processed MCMC chains for the mean parameters; if
-#' \code{y} is a matrix, a \eqn{true.iter \times 2 \times k} array with
+#' \code{y} is a matrix, a \eqn{true.iter \times D \times k} array with
 #' the post-processed MCMC chains for the mean parameters.}
 #' \item{\code{mcmc_sd}}{  If \code{y} is a vector, a \eqn{true.iter \times k}
 #' matrix with the post-processed MCMC chains for the sd parameters; if
-#' \code{y} is a matrix, a \eqn{true.iter \times 2} array with
+#' \code{y} is a matrix, a \eqn{true.iter \times D} array with
 #' the post-processed MCMC chains for the sd parameters.}
 #' \item{\code{mcmc_weight}}{A \eqn{true.iter \times k}
 #' matrix with the post-processed MCMC chains for the weights parameters.}
 #'\item{\code{mcmc_mean_raw}}{ If \code{y} is a vector, a \eqn{(nMC-burn) \times k} matrix
 #' with the raw MCMC chains for the mean parameters as given by JAGS; if
-#' \code{y} is a matrix, a \eqn{(nMC-burn) \times 2 \times k} array with the raw MCMC chains
+#' \code{y} is a matrix, a \eqn{(nMC-burn) \times D \times k} array with the raw MCMC chains
 #' for the mean parameters as given by JAGS/Stan.}
 #' \item{\code{mcmc_sd_raw}}{ If \code{y} is a vector, a \eqn{(nMC-burn) \times k} matrix
 #' with the raw MCMC chains for the sd parameters as given by JAGS/Stan; if
-#' \code{y} is a matrix, a \eqn{(nMC-burn) \times 2} array with the raw MCMC chains
+#' \code{y} is a matrix, a \eqn{(nMC-burn) \times D} array with the raw MCMC chains
 #' for the sd parameters as given by JAGS/Stan.}
 #' \item{\code{mcmc_weight_raw}}{A \eqn{(nMC-burn) \times k} matrix
 #' with the raw MCMC chains for the weights parameters as given by JAGS/Stan.}
@@ -185,6 +185,7 @@
 #' \code{"diana"} or \code{"hclust"}.}
 #' \item{\code{pivots}}{The vector of indices of pivotal units identified by the selected pivotal criterion.}
 #' \item{\code{model}}{The JAGS/Stan model code. Apply the \code{"cat"} function for a nice visualization of the code.}
+#' \item{\code{stanfit}}{An object of S4 class \code{stanfit} for the fitted model (only if \code{software="rstan"}).}
 #'
 #' @author Leonardo Egidi \email{legidi@units.it}
 #' @references Egidi, L., Pappadà, R., Pauli, F. and Torelli, N. (2018). Relabelling in Bayesian Mixture
@@ -196,15 +197,15 @@
 #'\dontrun{
 #' N   <- 200
 #' k   <- 4
+#' D   <- 2
 #' nMC <- 1000
-#' M1  <-c(-.5,8)
+#' M1  <- c(-.5,8)
 #' M2  <- c(25.5,.1)
 #' M3  <- c(49.5,8)
 #' M4  <- c(63.0,.1)
-#' Mu  <- matrix(rbind(M1,M2,M3,M4),c(4,2))
-#' sds <- cbind(rep(1,k), rep(20,k))
-#' Sigma.p1 <- matrix(c(sds[1,1]^2,0,0,sds[1,1]^2), nrow=2, ncol=2)
-#' Sigma.p2 <- matrix(c(sds[1,2]^2,0,0,sds[1,2]^2), nrow=2, ncol=2)
+#' Mu  <- rbind(M1,M2,M3,M4)
+#' Sigma.p1 <- diag(D)
+#' Sigma.p2 <- 20*diag(D)
 #' W <- c(0.2,0.8)
 #' sim <- piv_sim(N = N, k = k, Mu = Mu,
 #'                Sigma.p1 = Sigma.p1,
@@ -230,6 +231,7 @@
 #' ### Fishery data (bayesmix package)
 #'
 #'\dontrun{
+#' library(bayesmix)
 #' data(fish)
 #' y <- fish[,1]
 #' k <- 5
@@ -248,13 +250,14 @@
 
 
 
+
 piv_MCMC <- function(y,
                      k,
                      nMC,
                      priors,
-                     piv.criterion = "maxsumdiff",
-                     clustering = "diana",
-                     software = "rjags",
+                     piv.criterion = c("MUS", "maxsumint", "minsumnoint", "maxsumdiff"),
+                     clustering = c("diana", "hclust"),
+                     software = c("rjags", "rstan"),
                      burn =0.5*nMC,
                      chains = 4,
                      cores = 1){
@@ -262,30 +265,21 @@ piv_MCMC <- function(y,
   #### checks
 
   # piv.criterion
-  list_crit <- c("MUS", "maxsumint", "minsumnoint", "maxsumdiff")
-  if (sum(piv.criterion!=list_crit)==4){
-    stop(paste("object ", "'", piv.criterion,"'", " not found.
-    Please select one among the following pivotal
-    criteria: MUS, maxsumint, minsumnoint, maxsumdiff", sep=""))
+  if (missing(piv.criterion)){
+    piv.criterion <- "maxsumdiff"
   }
+  list_crit <- c("MUS", "maxsumint", "minsumnoint", "maxsumdiff")
+  piv.criterion <- match.arg(piv.criterion, list_crit)
 
   # clustering
 
   list_clust <- c("diana", "hclust")
-  if (sum(clustering!=list_clust)==2){
-    stop(paste("object ", "'", clustering,"'", " not found.
-    Please select one among the following
-    clustering methods: diana, hclust", sep=""))
-  }
+  clustering <- match.arg(clustering, list_clust)
 
   # software
 
   list_soft <- c("rjags", "rstan")
-  if (sum(software!=list_soft)==2){
-    stop(paste("object ", "'", software,"'", " not found.
-    Please select one among the following
-    softwares: rjags, rstan", sep=""))
-  }
+  software <- match.arg(software, list_soft)
 
   # burn-in
 
@@ -293,7 +287,6 @@ piv_MCMC <- function(y,
     stop("Please, 'burn' argument has to be minor than
          the number of MCMC iterations!", sep="")
   }
-
 
 
   ###
@@ -438,8 +431,8 @@ piv_MCMC <- function(y,
       if(missing(priors)){
         mu_0 <- 0
         B0inv <- 0.1
-        mu_phi <- 0
-        sigma_phi <- 2
+        mu_sigma <- 0
+        tau_sigma <- 2
       }else{
         if (is.null(priors$mu_0)){
           mu_0 <- 0
@@ -451,21 +444,21 @@ piv_MCMC <- function(y,
         }else{
           B0inv <- priors$B0inv
         }
-        if (is.null(priors$mu_phi)){
-          mu_phi <- 0
+        if (is.null(priors$mu_sigma)){
+          mu_sigma <- 0
         }else{
-          mu_phi <- priors$mu_phi
+          mu_sigma <- priors$mu_sigma
         }
-        if (is.null(priors$sigma_phi)){
-          sigma_phi <- 2
+        if (is.null(priors$tau_sigma)){
+          tau_sigma <- 2
         }else{
-          sigma_phi <- priors$sigma_phi
+          tau_sigma <- priors$tau_sigma
         }
       }
 
       data = list(N=N, y=y, k=k,
                   mu_0=mu_0, B0inv=B0inv,
-                  mu_phi=mu_phi, sigma_phi=sigma_phi)
+                  mu_sigma=mu_sigma, tau_sigma=tau_sigma)
       mix_univ <-"
         data {
           int<lower=1> k;          // number of mixture components
@@ -473,31 +466,31 @@ piv_MCMC <- function(y,
           real y[N];               // observations
           real mu_0;               // mean hyperparameter
           real<lower=0> B0inv;     // mean hyperprecision
-          real mu_phi;             // sigma hypermean
-          real<lower=0> sigma_phi; // sigma hyper sd
+          real mu_sigma;           // sigma hypermean
+          real<lower=0> tau_sigma; // sigma hyper sd
           }
         parameters {
-          simplex[k] theta;        // mixing proportions
+          simplex[k] eta;             // mixing proportions
           ordered[k] mu;              // locations of mixture components
           vector<lower=0>[k] sigma;   // scales of mixture components
           }
       transformed parameters{
-          vector[k] log_theta = log(theta);  // cache log calculation
+          vector[k] log_eta = log(eta);  // cache log calculation
           vector[k] pz[N];
           simplex[k] exp_pz[N];
               for (n in 1:N){
                   pz[n] =   normal_lpdf(y[n]|mu, sigma)+
-                            log_theta-
+                            log_eta-
                             log_sum_exp(normal_lpdf(y[n]|mu, sigma)+
-                            log_theta);
+                            log_eta);
                   exp_pz[n] = exp(pz[n]);
                             }
           }
       model {
-        sigma ~ lognormal(mu_phi, sigma_phi);
+        sigma ~ lognormal(mu_sigma, tau_sigma);
         mu ~ normal(mu_0, 1/B0inv);
             for (n in 1:N) {
-              vector[k] lps = log_theta;
+              vector[k] lps = log_eta;
                 for (j in 1:k){
                     lps[j] += normal_lpdf(y[n] | mu[j], sigma[j]);
                     target+=pz[n,j];
@@ -516,14 +509,15 @@ piv_MCMC <- function(y,
                         data=data,
                         chains =chains,
                         iter =nMC)
-      printed <- cat(print(fit_univ, pars =c("mu", "theta", "sigma")))
+      stanfit <- fit_univ
+      printed <- cat(print(fit_univ, pars =c("mu", "eta", "sigma")))
       sims_univ <- rstan::extract(fit_univ)
 
       J <- 3
-      mcmc.pars <- array(data = NA, dim = c(dim(sims_univ$theta)[1], k, J))
+      mcmc.pars <- array(data = NA, dim = c(dim(sims_univ$eta)[1], k, J))
       mcmc.pars[ , , 1] <- sims_univ$mu
       mcmc.pars[ , , 2] <- sims_univ$sigma
-      mcmc.pars[ , , 3] <- sims_univ$theta
+      mcmc.pars[ , , 3] <- sims_univ$eta
 
       mu_pre_switch_compl <-  mcmc.pars[ , , 1]
       tau_pre_switch_compl <-  mcmc.pars[ , , 2]
@@ -609,13 +603,15 @@ piv_MCMC <- function(y,
 
   }else if (is.matrix(y)){
     N <- dim(y)[1]
+    D <- dim(y)[2]
     # Parameters' initialization
     clust_inits <- KMeans(y, k)$cluster
     #cutree(hclust(dist(y), "average"),k)
-    mu_inits <- matrix(0,k,2)
+    mu_inits <- matrix(0,k,D)
     for (j in 1:k){
-      mu_inits[j,] <- cbind(mean(y[clust_inits==j,1]), mean(y[clust_inits==j,2]))
-    }
+      for (d in 1:D){
+      mu_inits[j, d] <- mean(y[clust_inits==j,d])
+    }}
     #Reorder mu_inits according to the x-coordinate
     mu_inits <-
       mu_inits[sort(mu_inits[,1], decreasing=FALSE, index.return=TRUE)$ix,]
@@ -626,23 +622,23 @@ piv_MCMC <- function(y,
 
       # Initial values
       if (missing(priors)){
-        mu_0 <- as.vector(c(0,0))
-        S2 <- matrix(c(1,0,0,1),nrow=2)/100000
-        S3 <- matrix(c(1,0,0,1),nrow=2)/100000
+        mu_0 <- rep(0, D)
+        S2 <- diag(D)/100000
+        S3 <- diag(D)/100000
         alpha <- rep(1,k)
       }else{
         if (is.null(priors$mu_0)){
-          mu_0 <- as.vector(c(0,0))
+          mu_0 <- rep(0, D)
         }else{
           mu_0 <- priors$mu_0
         }
         if (is.null(priors$S2)){
-          S2 <- matrix(c(1,0,0,1),nrow=2)/100000
+          S2 <- diag(D)/100000
         }else{
           S2 <- priors$S2
         }
         if (is.null(priors$S3)){
-          S3 <- matrix(c(1,0,0,1),nrow=2)/100000
+          S3 <- diag(D)/100000
         }else{
           S3 <- priors$S3
         }
@@ -654,7 +650,7 @@ piv_MCMC <- function(y,
       }
 
       # Data
-      dati.biv <- list(y = y, N = N, k = k,
+      dati.biv <- list(y = y, N = N, k = k, D = D,
                        S2= S2, S3= S3, mu_0=mu_0,
                        alpha = alpha)
 
@@ -663,26 +659,27 @@ piv_MCMC <- function(y,
     # Likelihood:
 
     for (i in 1:N){
-      yprev[i,1:2]<-y[i,1:2]
-      y[i,1:2] ~ dmnorm(muOfClust[clust[i],],tauOfClust)
-      clust[i] ~ dcat(pClust[1:k] )
+      yprev[i,1:D]<-y[i,1:D]
+      y[i,1:D] ~ dmnorm(mu[clust[i],],tau)
+      clust[i] ~ dcat(eta[1:k] )
     }
 
     # Prior:
 
     for (g in 1:k) {
-      muOfClust[g,1:2] ~ dmnorm(mu_0[],S2[,])}
-      tauOfClust[1:2,1:2] ~ dwish(S3[,],3)
-      Sigma[1:2,1:2] <- inverse(tauOfClust[,])
-      pClust[1:k] ~ ddirch(alpha)
+      mu[g,1:D] ~ dmnorm(mu_0[],S2[,])}
+      tau[1:D,1:D] ~ dwish(S3[,],D+1)
+      Sigma[1:D,1:D] <- inverse(tau[,])
+      eta[1:k] ~ ddirch(alpha)
   }"
 
       init1.biv <- list()
-      for (s in 1:chains)
-      init1.biv[[s]] <- dump.format(list(muOfClust=mu_inits,
-                                    tauOfClust= matrix(c(15,0,0,15),ncol=2),
-                                    pClust=rep(1/k,k), clust=clust_inits))
-      moni.biv <- c("clust","muOfClust","tauOfClust","pClust")
+      for (s in 1:chains){
+      init1.biv[[s]] <- dump.format(list(mu=mu_inits,
+                                    tau= 15*diag(D),
+                                    eta=rep(1/k,k), clust=clust_inits))
+      }
+      moni.biv <- c("clust","mu","tau","eta")
 
       mod   <- mod.mist.biv
       dati  <- dati.biv
@@ -693,9 +690,9 @@ piv_MCMC <- function(y,
       ogg.jags <- run.jags(model=mod, data=dati, monitor=moni,
                            inits=init1, n.chains=chains,plots=FALSE, thin=1,
                            sample=nMC, burnin=burn)
-      printed <- print(add.summary(ogg.jags, vars= c("muOfClust",
-                      "tauOfClust",
-                      "pClust")))
+      printed <- print(add.summary(ogg.jags, vars= c("mu",
+                      "tau",
+                      "eta")))
       # Extraction
       ris <- ogg.jags$mcmc[[1]]
 
@@ -703,14 +700,17 @@ piv_MCMC <- function(y,
       group <- ris[-(1:burn),grep("clust[",colnames(ris),fixed=TRUE)]
 
       # only the variances
-      tau <- sqrt( (1/ris[-(1:burn),grep("tauOfClust[",colnames(ris),fixed=TRUE)])[,c(1,4)])
-      prob.st <- ris[-(1:burn),grep("pClust[",colnames(ris),fixed=TRUE)]
+      tau <- sqrt( (1/ris[-(1:burn),grep("tau[",colnames(ris),fixed=TRUE)])[,c(1,4)])
+      prob.st <- ris[-(1:burn),grep("eta[",colnames(ris),fixed=TRUE)]
       M <- nrow(group)
       H <- list()
 
-      mu_pre_switch_compl <- array(rep(0, M*2*k), dim=c(M,2,k))
+      mu_pre_switch_compl <- array(rep(0, M*D*k), dim=c(M,D,k))
       for (i in 1:k){
-        H[[i]] <- ris[-(1:burn),grep("muOfClust",colnames(ris),fixed=TRUE)][,c(i,i+k)]
+        H[[i]] <- ris[-(1:burn),
+                      grep(paste("mu[",i, sep=""),
+                           colnames(ris),fixed=TRUE)]
+                         #[,c(i,i+k)]
       }
       for (i in 1:k){
         mu_pre_switch_compl[,,i] <- as.matrix(H[[i]])
@@ -725,9 +725,11 @@ piv_MCMC <- function(y,
         #return(1)
       }else{
         L<-list()
-        mu_pre_switch <- array(rep(0, true.iter*2*k), dim=c(true.iter,2,k))
+        mu_pre_switch <- array(rep(0, true.iter*D*k), dim=c(true.iter,D,k))
         for (i in 1:k){
-          L[[i]] <- ris[,grep("muOfClust",colnames(ris),fixed=TRUE)][,c(i,i+k)]
+          L[[i]] <- ris[,grep(paste("mu[", i, sep=""),
+                                    colnames(ris),fixed=TRUE)]
+          #[,c(i,i+k)]
         }
         for (i in 1:k){
           mu_pre_switch[,,i] <- as.matrix(L[[i]])
@@ -742,26 +744,26 @@ piv_MCMC <- function(y,
       mcmc_weight_raw = prob.st
       mcmc_sd_raw = tau
 
-      tau <- sqrt( (1/ris[,grep("tauOfClust[",colnames(ris),fixed=TRUE)])[,c(1,4)])
-      prob.st <- ris[,grep("pClust[",colnames(ris),fixed=TRUE)]
+      tau <- sqrt( (1/ris[,grep("tau[",colnames(ris),fixed=TRUE)])[,c(1,4)])
+      prob.st <- ris[,grep("eta[",colnames(ris),fixed=TRUE)]
       mu <- mu_pre_switch
 
 
     }else if(software=="rstan"){
       if (missing(priors)){
-        mu_0 <- c(0,0)
-        eta <- 1
+        mu_0 <- rep(0, D)
+        epsilon <- 1
         sigma_d <- 2.5
       }else{
         if (is.null(priors$mu_0)){
-          mu_0 <- c(0,0)
+          mu_0 <- rep(0, D)
         }else{
           mu_0 <- priors$mu_0
         }
-        if (is.null(priors$eta)){
-          eta <- 1
+        if (is.null(priors$epsilon)){
+          epsilon <- 1
         }else{
-          eta  <- priors$eta
+          epsilon  <- priors$epsilon
         }
         if (is.null(priors$sigma_d)){
           sigma_d <- 2.5
@@ -769,8 +771,8 @@ piv_MCMC <- function(y,
           sigma_d <- priors$sigma_d
         }
       }
-      data =list(N=N, k=k, y=y, D=2, mu_0=mu_0,
-                 eta = eta, sigma_d = sigma_d)
+      data =list(N=N, k=k, y=y, D=D, mu_0=mu_0,
+                 epsilon = epsilon, sigma_d = sigma_d)
       mix_biv <- "
         data {
           int<lower=1> k;          // number of mixture components
@@ -778,11 +780,11 @@ piv_MCMC <- function(y,
           int D;                   // data dimension
           matrix[N,D] y;           // observations matrix
           vector[D] mu_0;
-          real<lower=0> eta;
+          real<lower=0> epsilon;
           real<lower=0> sigma_d;
         }
         parameters {
-          simplex[k] theta;        // mixing proportions
+          simplex[k] eta;         // mixing proportions
           vector[D] mu[k];        // locations of mixture components
           cholesky_factor_corr[D] L_Omega;   // scales of mixture components
           vector<lower=0>[D] L_sigma;
@@ -790,7 +792,7 @@ piv_MCMC <- function(y,
           vector<lower=0>[D] L_tau;
           }
         transformed parameters{
-          vector[k] log_theta = log(theta);  // cache log calculation
+          vector[k] log_eta = log(eta);  // cache log calculation
           vector[k] pz[N];
           simplex[k] exp_pz[N];
           matrix[D,D] L_Sigma=diag_pre_multiply(L_sigma, L_Omega);
@@ -799,19 +801,19 @@ piv_MCMC <- function(y,
 
             for (n in 1:N){
                 pz[n]=   multi_normal_cholesky_lpdf(y[n]|mu, L_Sigma)+
-                         log_theta-
+                         log_eta-
                          log_sum_exp(multi_normal_cholesky_lpdf(y[n]|
                                                      mu, L_Sigma)+
-                         log_theta);
+                         log_eta);
                 exp_pz[n] = exp(pz[n]);
               }
           }
         model{
-          L_Omega ~ lkj_corr_cholesky(eta);
+          L_Omega ~ lkj_corr_cholesky(epsilon);
           L_sigma ~ cauchy(0, sigma_d);
           mu ~ multi_normal_cholesky(mu_0, L_Tau);
             for (n in 1:N) {
-              vector[k] lps = log_theta;
+              vector[k] lps = log_eta;
                 for (j in 1:k){
                     lps[j] += multi_normal_cholesky_lpdf(y[n] |
                                                    mu[j], L_Sigma);
@@ -831,7 +833,8 @@ piv_MCMC <- function(y,
                        data=data,
                        chains =chains,
                        iter =nMC)
-      printed <- cat(print(fit_biv, pars=c("mu", "theta", "L_Sigma")))
+      stanfit <- fit_biv
+      printed <- cat(print(fit_biv, pars=c("mu", "eta", "L_Sigma")))
       sims_biv <- rstan::extract(fit_biv)
 
       # Extraction
@@ -840,10 +843,10 @@ piv_MCMC <- function(y,
       # Post- process of the chains----------------------
       group <- sims_biv$z
       tau <- sims_biv$L_sigma
-      prob.st <- sims_biv$theta
+      prob.st <- sims_biv$eta
       M <- nrow(group)
 
-      mu_pre_switch_compl <- array(rep(0, M*2*k), dim=c(M,2,k))
+      mu_pre_switch_compl <- array(rep(0, M*D*k), dim=c(M,D,k))
       for (i in 1:M)
         mu_pre_switch_compl[i,,] <- t(sims_biv$mu[i,,])
       # Discard iterations
@@ -856,7 +859,7 @@ piv_MCMC <- function(y,
         #return(1)
       }else{
 
-        mu_pre_switch <- array(rep(0, true.iter*2*k), dim=c(true.iter,2,k))
+        mu_pre_switch <- array(rep(0, true.iter*D*k), dim=c(true.iter,D,k))
         for (i in 1:true.iter)
           mu_pre_switch[i,,] <- t(sm[i,,])
       }
@@ -868,7 +871,7 @@ piv_MCMC <- function(y,
       group <- sims_biv$z[numeffettivogruppi==k,]
       mu <- mu_pre_switch
       tau <- sims_biv$L_sigma[numeffettivogruppi==k, ]
-      prob.st <- sims_biv$theta[numeffettivogruppi==k,]
+      prob.st <- sims_biv$eta[numeffettivogruppi==k,]
       FreqGruppiJags <- table(group)
 
       model_code <- mix_biv
@@ -892,7 +895,7 @@ piv_MCMC <- function(y,
     if (cont > 1){
       k <- cont
     }
-    mu_switch  <- array(rep(0, true.iter*2*k), dim=c(true.iter,2,k))
+    mu_switch  <- array(rep(0, true.iter*D*k), dim=c(true.iter,D,k))
     prob.st_switch <-  array(0, dim=c(true.iter,k))
     group <- group*0
     z <- array(0,dim=c(N, k, true.iter))
@@ -977,7 +980,7 @@ piv_MCMC <- function(y,
     if (k <=4 & sum(C==0)!=0){
 
       mus_res    <- MUS(C, grr)
-      clust  <-  mus_res$pivots
+      pivots     <-  mus_res$pivots
 
     }else{
 
@@ -986,6 +989,10 @@ piv_MCMC <- function(y,
       clust  <-  piv_sel(C=C,  clusters=as.vector(grr))
       pivots <- clust$pivots[,3]
     }
+  }
+
+  if (software == "rjags"){
+    stanfit = NULL
   }
 
 
@@ -1005,5 +1012,7 @@ piv_MCMC <- function(y,
                grr=grr,
                pivots = pivots,
                #print = printed,
-               model = model_code))
+               model = model_code,
+               k = k,
+               stanfit = stanfit))
 }