+++ inst/doc/Relabelling_in_Bayesian_mixtures_by_pivotal_units.R
@@ -6,23 +6,23 @@ knitr::opts_chunk$set(
 
 ## ----load, warning =FALSE, message = FALSE-------------------------------
 library(pivmet)
+library(bayesmix)
+library(bayesplot)
 
 ## ----nested, fig.align ='center'-----------------------------------------
-set.seed(50)
+set.seed(500)
 N  <- 200
 k  <- 3
+D <- 2
 nMC <- 2000
 M1 <- c(-10,8)
 M2 <- c(10,.1)
 M3 <- c(30,8)
 # matrix of input means
-Mu <- matrix(rbind(M1,M2,M3),c(k,2)) 
-sds    <- cbind(rep(1,k), rep(15,k))
+Mu <- rbind(M1,M2,M3)
 # covariance matrices for the two subgroups
-Sigma.p1 <- matrix(c(sds[1,1]^2,0,0,sds[1,1]^2),
-nrow=2, ncol=2)
-Sigma.p2 <- matrix(c(sds[1,2]^2,0,0,sds[1,2]^2),
-nrow=2, ncol=2)
+Sigma.p1 <- diag(D)
+Sigma.p2 <- (10^2)*diag(D)
 # subgroups' weights
 W   <- c(0.2,0.8)
 # simulate data
@@ -30,7 +30,8 @@ sim <- piv_sim(N = N, k = k, Mu = Mu,
  Sigma.p1 = Sigma.p1, Sigma.p2 = Sigma.p2, W = W)
 
 ## ----mcmc, message =  FALSE, warning = FALSE-----------------------------
-res <- piv_MCMC(y = sim$y, k= k, nMC =nMC)
+res <- piv_MCMC(y = sim$y, k= k, nMC =nMC, 
+                piv.criterion = "maxsumdiff")
 
 ## ----pivotal_rel, fig.show='hold', fig.align='center',fig.width=7--------
 rel <- piv_rel(mcmc=res)
@@ -48,8 +49,8 @@ hist(y, breaks=40, prob = TRUE, cex.lab=1.6,
 ## ----fish_data-----------------------------------------------------------
 k <- 5
 nMC <- 15000
-res <- piv_MCMC(y = y, k = k, nMC = nMC, burn = 0.5*nMC,
-software = "rjags")
+res <- piv_MCMC(y = y, k = k, nMC = nMC, 
+                burn = 0.5*nMC, software = "rjags")
 
 ## ----true_iter-----------------------------------------------------------
 res$true.iter
@@ -58,6 +59,18 @@ res$true.iter
 rel <- piv_rel(mcmc=res)
 piv_plot(y=y, res, rel, par = "mean", type="chains")
 piv_plot(y=y, res, rel, type="hist")
+
+## ----stan, eval = FALSE, fig.align= 'center', fig.width=7----------------
+#  # stan code not evaluated here
+#  res2 <- piv_MCMC(y = y, k = k, nMC = 3000,
+#                   software = "rstan")
+#  rel2 <- piv_rel(res2)
+#  piv_plot(y=y, res2, rel2, par = "mean", type="chains")
+
+## ----bayesplot, eval = FALSE, fig.align= 'center', fig.width=5-----------
+#  # stan code not evaluated here
+#  posterior <- as.array(res2$stanfit)
+#  mcmc_intervals(posterior, regex_pars = c("mu"))
 
 ## ----model_code----------------------------------------------------------
 cat(res$model)